OGCN- BASE/backend/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Flask-CORS==3.0.10
requests==2.28.2
ollama==0.1.6
pillow==10.1.0
cerebras.cloud.sdk
openai
